project/requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
yfinance
beautifulsoup4
requests
python-dotenv
flask
pickle-mixin
pyarrow
fastparquet
scipy.stats